fully implemented matching
given a complete set of rankings for each user, number of users, and  preferred number of people in each group, program will generate groups using Gale Shipley Algorithm making sure everyone is matched. Generates 200 groups of 4 (800 total which is max class size at Stanford) in about 5 seconds so time is not an issue.